+++ CHANGELOG.md
@@ -29,6 +29,7 @@ Updates should follow the [Keep a CHANGELOG](https://keepachangelog.com/) princi
  - `Node::replaceChildren()` now accepts any type of `iterable`, not just `array`s
  - Some block elements now extend `AbstractStringContainerBlock` instead of `AbstractBlock`
  - `InlineContainerInterface` now extends the new `StringContainerInterface`
+ - The `handleRemainingContents()` method (formerly on `AbstractBlock`, now on `AbstractStringContainerBlock`) is now an `abstract method
 
 ### Removed